refactor(rust-api): extract lora catalog/import/validation into loras.rs (sc-1639 Phase 1, slice 3)
Third behavior-preserving slice of the sc-1639 lib.rs Value-cluster split.
Moves the lora-exclusive code out of lib.rs into a new loras.rs:
- handlers: list_loras, delete_lora, lora_catalog, create_lora_import_job +
  import plumbing (queue/from_multipart/write_upload/cleanup/max_upload_bytes)
- normalize: normalize_lora_entry
- lora-spec validation: validate_job_lora_compatibility, validate_lora_specs_for_model,
  hydrate_lora_spec, validate_lora_safetensors_header, read_safetensors_header_for_api
- import source/family: lora_source_provider, lora_url_error_message,
  detect_family_from_local_path, reconcile_lora_family
- uploads sweep: sweep_stale_lora_uploads(+_before)
- artifact/install: lora_is_installed, lora_artifact_paths,
  lora_huggingface_cached_file, lora_families

Family validators shared with model import + training (validate_lora_family,
normalize_lora_family, known_lora_families(+_from_manifests), model_lora_families,
families_from_value_chain) and generic upload/url util (validate_source_url,
sanitized_upload_filename, validate_lora_import_source_path, slugify_lora_id,
lora_source_url_file_stem) stay at root; loras.rs reaches them via `use super::*`.
Same sc-1636 pattern: pub(crate) on the external surface + `use loras::*`
re-export so lib.rs/generation.rs/characters.rs/jobs.rs/training.rs/tests.rs
call-sites resolve unchanged.

lib.rs 3,945 -> 2,990. Verified the only lib.rs change vs main is the two-line
mod declaration. cargo test (62 rust-api), npm run rust:check, and
cargo build --features embed-web all green.
